docs: some doc fixes and improvements
